% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwbGS_example.R
\docType{data}
\name{ad}
\alias{ad}
\title{\dontrun{}
\format{
An object of class \code{list} (inherits from \code{curtailment_single}) of length 2.
}
\usage{
ad
}
\description{
\dontrun{
}
\keyword{datasets}
