% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empSE.R
\name{empSE}
\alias{empSE}
\title{Empirical standard error}
\usage{
empSE(theta.hat)
}
\arguments{
\item{theta.hat}{vector estimated response rates}
}
\value{
Estimated empirical standard error
}
\description{
Finds empirical standard error for a vector of estimates.
}
