% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maGS.R
\name{maGS}
\alias{maGS}
\title{maGS}
\usage{
maGS(theta = 0.5, des, bounds, nsims = 1e+05, n.studies = 4)
}
\arguments{
\item{theta}{true response probablity}

\item{des}{a realisation of a group sequential design, formatted as per
output of curtailment::singlearmDesign}

\item{bounds}{interim analysis bounds, also from curtailment::singlearmDesign}

\item{nsims}{number of simulations (default 1e5)}

\item{n.studies}{number of non-adaptive studies in meta-analysis}
}
\value{
Data frame of results
}
\description{
Examines bias for a single group sequential design WRT a meta-analysis of non-adaptive designs
}
