% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maSimon.R
\name{maSimon}
\alias{maSimon}
\title{maSimon}
\usage{
maSimon(theta = 0.5, nsims = 1e+05, n.studies = 4, des)
}
\arguments{
\item{theta}{true response probablity}

\item{nsims}{number of simulations (default 1e5)}

\item{n.studies}{number of non-adaptive studies in meta-analysis}

\item{des}{a realisation of a Simon design, formatted as per output of clinfun::ph2simon}
}
\value{
Data frame of results
}
\description{
Examines bias for a single Simon design WRT a meta-analysis of non-adaptive designs
}
