% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCSEbias.R
\name{MCSEbias}
\alias{MCSEbias}
\title{Monte Carlo SE of bias}
\usage{
MCSEbias(theta.hat)
}
\arguments{
\item{theta.hat}{vector estimated response rates}
}
\value{
Monte Carlo standard error of bias
}
\description{
Finds Monte Carlo standard error of bias for a vector of estimates.
}
