% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwb2stage.R
\name{pwb2stage}
\alias{pwb2stage}
\title{pwb2stage}
\usage{
pwb2stage(N = 2428, theta0 = 0.2, theta1 = 0.15, nsims = 10000)
}
\arguments{
\item{N}{Maximum sample size}

\item{theta0}{true response probablity for control arm}

\item{theta1}{true response probablity for treatment arm}

\item{nsims}{number of simulations (default 1e4)}
}
\value{
list of length 2:
ests: estimates
mc.error
}
\description{
For two-arm two-stage design, obtains bias for a single true response rate.
}
