% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwbSimon.R
\name{pwbSimon}
\alias{pwbSimon}
\title{pwbSimon}
\usage{
pwbSimon(theta, des, nsims = 1e+05)
}
\arguments{
\item{theta}{true response probablity}

\item{des}{a realisation of a Simon design, formatted as per output of clinfun::ph2simon}

\item{nsims}{number of simulations (default 1e5)}
}
\value{
list of length 3:
results: individual sims
ests: estimates
mc.error
}
\description{
Examines bias for a single Simon design
}
