% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showTable.R
\name{showTable}
\alias{showTable}
\title{showTable}
\usage{
showTable(bias.df, theta, latex = FALSE, digit = 4, inc.type = FALSE)
}
\arguments{
\item{bias.df}{The data frame containing all results}

\item{theta}{Response rate}

\item{latex}{Print Latex code for table (TRUE/FALSE)}

\item{digit}{Number of rounding digits for bias and SE}

\item{inc.type}{Include column for type (Default: FALSE)}
}
\value{
A data frame showing the results for the single theta
}
\description{
Shows results for single value of theta
}
