% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMeta.R
\name{simMeta}
\alias{simMeta}
\title{Meta-analysis simulation (binary outcome)}
\usage{
simMeta(N, theta, n.studies, nsims)
}
\arguments{
\item{N}{sample size (shared for all trials)}

\item{theta}{response rate (shared for all trials)}

\item{n.studies}{number of trials to simulate}

\item{nsims}{number of meta-analyses to simulate}
}
\value{
A list of two vectors of length N: response rate theta hat and SE
}
\description{
Simulates binary data for a meta-analysis.
}
